requirements.txt
# Voice Assistant Backend Dependencies
flask>=2.3.0
flask-cors>=4.0.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
openai-whisper>=20231117
openai>=1.0.0
TTS>=0.16.6
transformers==4.49.0

# Additional dependencies for audio processing
librosa>=0.10.0
# pyaudio already installed via brew + pip

# For better error handling and logging
werkzeug>=2.3.0
gunicorn>=21.0.0